shrink type for subgroup size

+++ include/oneapi/dpl/pstl/hetero/dpcpp/parallel_backend_sycl_reduce_then_scan.h
@@ -1284,7 +1284,7 @@ inline std::uint32_t
 __get_reduce_then_scan_workgroup_size(const sycl::queue& q)
 {
     const std::size_t __max_wg_size = oneapi::dpl::__internal::__max_work_group_size(q);
-    const std::size_t __sg_size = __get_reduce_then_scan_workaround_sg_sz();
+    const std::uint8_t __sg_size = __get_reduce_then_scan_workaround_sg_sz();
     // Round down to multiple of sub-group size
     return __max_wg_size  - (__max_wg_size % __sg_size);
 }
